Moved /register contents into a util file (#431)
